+++ cita-chain/types/src/lib.rs
@@ -31,11 +31,10 @@ pub mod block;
 pub mod block_number;
 pub mod block_receipts;
 pub mod context;
+pub mod db_indexes;
 pub mod errors;
-pub mod extras;
 pub mod filter;
 pub mod header;
-pub mod indexes;
 pub mod log_blooms;
 pub mod log_entry;
 pub mod receipt;
